+++ Runtime/ARSubsystems/CameraSubsystem/LightEstimationMode.cs
@@ -1,10 +1,12 @@
+using System;
 using System.ComponentModel;
 
 namespace UnityEngine.XR.ARSubsystems
 {
     /// <summary>
-    /// Represents the light estimation mode. This is deprecated.
+    /// Represents the light estimation mode.
     /// </summary>
+    [Obsolete("This enum is no longer used. Its functionality was replaced by the Feature flags enum. (2023-02-02)")]
     public enum LightEstimationMode
     {
         /// <summary>